KAN-18 Add Logic of Update Booking

+++ Wasla_Backend/Services/Implementation/DoctorServiceService.cs
@@ -56,6 +56,9 @@
             if (service == null)
                 throw new NotFoundException("ServiceNotFound");
 
+            if (service.ServiceDays.Any(s => s.isBooking))
+                throw new BadRequestException("CannotUpdateServiceWithExistingBookings");
+
             service.serviceName = dto.serviceName;
             service.description = dto.description;
             service.price = dto.price;
@@ -128,9 +131,10 @@
             var service =  await _doctorServiceRepository.GetByIdAsync(serviceId);
             
             if(service == null)
-            {
                 throw new NotFoundException("ServiceNotFound");
-            }
+
+            if (service.ServiceDays.Any(s => s.isBooking))
+                throw new BadRequestException("CannotUpdateServiceWithExistingBookings");
 
             await _doctorServiceRepository.DeleteByIdAsync(serviceId);
             await _doctorServiceRepository.SaveChangesAsync();
